+++ kmock-processor/src/test/resources/factory/expected/interfaze/SharedActual.kt
@@ -8,6 +8,7 @@ import kotlin.Any
 import kotlin.Boolean
 import kotlin.Comparable
 import kotlin.Suppress
+import kotlin.reflect.KClass
 import tech.antibytes.kmock.KMockContract
 import tech.antibytes.kmock.KMockContract.Collector
 
@@ -59,7 +60,7 @@ private inline fun <reified Mock : SpyOn, reified SpyOn : Shared1<K, L>, K : Any
     relaxed: Boolean,
     relaxUnitFun: Boolean,
     freeze: Boolean,
-    templateType: kotlin.reflect.KClass<factory.template.interfaze.Shared1<*, *>>,
+    templateType: KClass<Shared1<*, *>>,
 ): Mock where L : Any, L : Comparable<L> = when (Mock::class) {
     factory.template.interfaze.Shared1Mock::class -> factory.template.interfaze.Shared1Mock<K,
         L>(collector = collector, relaxUnitFun = relaxUnitFun, freeze = freeze, spyOn = spyOn as
@@ -72,7 +73,7 @@ internal actual inline fun <reified Mock : Shared1<K, L>, K : Any, L> kmock(
     relaxed: Boolean,
     relaxUnitFun: Boolean,
     freeze: Boolean,
-    templateType: kotlin.reflect.KClass<factory.template.interfaze.Shared1<*, *>>,
+    templateType: KClass<Shared1<*, *>>,
 ): Mock where L : Any, L : Comparable<L> = getMockInstance(
     spyOn = null,
     collector = collector,
@@ -86,7 +87,7 @@ internal actual inline fun <reified Mock : SpyOn, reified SpyOn : Shared1<K, L>,
     spyOn: SpyOn,
     collector: KMockContract.Collector,
     freeze: Boolean,
-    templateType: kotlin.reflect.KClass<factory.template.interfaze.Shared1<*, *>>,
+    templateType: KClass<Shared1<*, *>>,
 ): Mock where L : Any, L : Comparable<L> = getMockInstance(
     spyOn = spyOn,
     collector = collector,
@@ -103,7 +104,7 @@ private inline fun <reified Mock : SpyOn, reified SpyOn : Shared.Shared4<K, L>, 
     relaxed: Boolean,
     relaxUnitFun: Boolean,
     freeze: Boolean,
-    templateType: kotlin.reflect.KClass<factory.template.interfaze.Shared.Shared4<*, *>>,
+    templateType: KClass<Shared.Shared4<*, *>>,
 ): Mock where L : Any, L : Comparable<L> = when (Mock::class) {
     factory.template.interfaze.Shared4Mock::class -> factory.template.interfaze.Shared4Mock<K,
         L>(collector = collector, relaxUnitFun = relaxUnitFun, freeze = freeze, spyOn = spyOn as
@@ -116,7 +117,7 @@ internal actual inline fun <reified Mock : Shared.Shared4<K, L>, K : Any, L> kmo
     relaxed: Boolean,
     relaxUnitFun: Boolean,
     freeze: Boolean,
-    templateType: kotlin.reflect.KClass<factory.template.interfaze.Shared.Shared4<*, *>>,
+    templateType: KClass<Shared.Shared4<*, *>>,
 ): Mock where L : Any, L : Comparable<L> = getMockInstance(
     spyOn = null,
     collector = collector,