+++ kmock-processor/src/test/resources/factory/expected/spiesonly/Platform.kt
@@ -7,6 +7,7 @@ import kotlin.Any
 import kotlin.Boolean
 import kotlin.Comparable
 import kotlin.Suppress
+import kotlin.reflect.KClass
 import tech.antibytes.kmock.KMockContract
 import tech.antibytes.kmock.KMockContract.Collector
 import tech.antibytes.kmock.proxy.NoopCollector
@@ -46,7 +47,7 @@ private inline fun <reified Mock : SpyOn, reified SpyOn : Platform1<K, L>, K : A
     relaxed: Boolean,
     relaxUnitFun: Boolean,
     freeze: Boolean,
-    templateType: kotlin.reflect.KClass<factory.template.spiesonly.Platform1<*, *>>,
+    templateType: KClass<Platform1<*, *>>,
 ): Mock where L : Any, L : Comparable<L> = when (Mock::class) {
     factory.template.spiesonly.Platform1Mock::class -> factory.template.spiesonly.Platform1Mock<K,
         L>(collector = collector, relaxUnitFun = relaxUnitFun, freeze = freeze, spyOn = spyOn as
@@ -58,7 +59,7 @@ internal inline fun <reified Mock : SpyOn, reified SpyOn : Platform1<K, L>, K : 
     spyOn: SpyOn,
     collector: KMockContract.Collector = NoopCollector,
     freeze: Boolean = true,
-    templateType: kotlin.reflect.KClass<factory.template.spiesonly.Platform1<*, *>>,
+    templateType: KClass<Platform1<*, *>>,
 ): Mock where L : Any, L : Comparable<L> = getMockInstance(
     spyOn = spyOn,
     collector = collector,