+++ kmock-processor/src/test/resources/factory/expected/spiesonly/SharedExpect.kt
@@ -7,6 +7,7 @@ import kotlin.Any
 import kotlin.Boolean
 import kotlin.Comparable
 import kotlin.Suppress
+import kotlin.reflect.KClass
 import tech.antibytes.kmock.KMockContract
 import tech.antibytes.kmock.KMockContract.Collector
 import tech.antibytes.kmock.proxy.NoopCollector
@@ -15,5 +16,5 @@ internal expect inline fun <reified Mock : SpyOn, reified SpyOn : Shared1<K, L>,
     spyOn: SpyOn,
     collector: KMockContract.Collector = NoopCollector,
     freeze: Boolean = true,
-    templateType: kotlin.reflect.KClass<factory.template.spiesonly.Shared1<*, *>>,
+    templateType: KClass<Shared1<*, *>>,
 ): Mock where L : Any, L : Comparable<L>