+++ kmock-processor/src/test/resources/factory/expected/spy/CommonActual.kt
@@ -7,6 +7,7 @@ import kotlin.Any
 import kotlin.Boolean
 import kotlin.Comparable
 import kotlin.Suppress
+import kotlin.reflect.KClass
 import tech.antibytes.kmock.KMockContract
 import tech.antibytes.kmock.KMockContract.Collector
 
@@ -55,7 +56,7 @@ private inline fun <reified Mock : SpyOn, reified SpyOn : Common1<K, L>, K : Any
     relaxed: Boolean,
     relaxUnitFun: Boolean,
     freeze: Boolean,
-    templateType: kotlin.reflect.KClass<factory.template.spy.Common1<*, *>>,
+    templateType: KClass<Common1<*, *>>,
 ): Mock where L : Any, L : Comparable<L> = when (Mock::class) {
     factory.template.spy.Common1Mock::class -> factory.template.spy.Common1Mock<K, L>(collector =
     collector, relaxUnitFun = relaxUnitFun, freeze = freeze, spyOn = spyOn as
@@ -68,7 +69,7 @@ internal actual inline fun <reified Mock : Common1<K, L>, K : Any, L> kmock(
     relaxed: Boolean,
     relaxUnitFun: Boolean,
     freeze: Boolean,
-    templateType: kotlin.reflect.KClass<factory.template.spy.Common1<*, *>>,
+    templateType: KClass<Common1<*, *>>,
 ): Mock where L : Any, L : Comparable<L> = getMockInstance(
     spyOn = null,
     collector = collector,
@@ -82,7 +83,7 @@ internal actual inline fun <reified Mock : SpyOn, reified SpyOn : Common1<K, L>,
     spyOn: SpyOn,
     collector: KMockContract.Collector,
     freeze: Boolean,
-    templateType: kotlin.reflect.KClass<factory.template.spy.Common1<*, *>>,
+    templateType: KClass<Common1<*, *>>,
 ): Mock where L : Any, L : Comparable<L> = getMockInstance(
     spyOn = spyOn,
     collector = collector,