+++ kmock-processor/src/test/resources/multi/expected/platformGeneric/GenericActualFactory.kt
@@ -6,6 +6,7 @@ import kotlin.Any
 import kotlin.Boolean
 import kotlin.Comparable
 import kotlin.Suppress
+import kotlin.reflect.KClass
 import multi.template.platformGeneric.Generic1
 import multi.template.platformGeneric.GenericPlatformContract
 import multi.template.platformGeneric.nested.Generic2
@@ -44,46 +45,36 @@ private inline fun <reified Mock : SpyOn, reified SpyOn, KMockTypeParameter0 : A
     relaxed: Boolean,
     relaxUnitFun: Boolean,
     freeze: Boolean,
-    templateType0: kotlin.reflect.KClass<multi.template.platformGeneric.Generic1<KMockTypeParameter0,
-        KMockTypeParameter1>>,
-    templateType1: kotlin.reflect.KClass<multi.template.platformGeneric.nested.Generic2<KMockTypeParameter2,
-        KMockTypeParameter3>>,
-    templateType2: kotlin.reflect.KClass<multi.template.platformGeneric.GenericPlatformContract.Generic3<KMockTypeParameter4,
-        KMockTypeParameter5>>,
+    templateType0: KClass<Generic1<*, *>>,
+    templateType1: KClass<Generic2<*, *>>,
+    templateType2: KClass<GenericPlatformContract.Generic3<*, *>>,
 ): Mock where SpyOn : Generic1<KMockTypeParameter0, KMockTypeParameter1>, SpyOn :
 Generic2<KMockTypeParameter2, KMockTypeParameter3>, SpyOn :
               GenericPlatformContract.Generic3<KMockTypeParameter4, KMockTypeParameter5>, KMockTypeParameter1
               : Any, KMockTypeParameter1 : Comparable<KMockTypeParameter1>, KMockTypeParameter3 : Any,
-              KMockTypeParameter3 : Comparable<KMockTypeParameter3> = if (Mock::class ==
-    multi.PlatformGenericMultiMock::class) {
-    multi.PlatformGenericMultiMock(collector = collector, freeze = freeze, spyOn = spyOn as SpyOn?) as
-        Mock
-} else {
-    throw RuntimeException("Unknown Interface ${Mock::class.simpleName}.")}
+              KMockTypeParameter3 : Comparable<KMockTypeParameter3> = multi.PlatformGenericMultiMock(collector
+= collector, freeze = freeze, spyOn = spyOn as SpyOn?) as Mock
 
-internal inline fun <reified Mock, KMockTypeParameter0 : Any, KMockTypeParameter1,
-    KMockTypeParameter2 : Any, KMockTypeParameter3, KMockTypeParameter4, KMockTypeParameter5> kmock(
+internal inline fun <reified Mock :
+PlatformGenericMultiMock<KMockTypeParameter0, KMockTypeParameter1, KMockTypeParameter2, KMockTypeParameter3, KMockTypeParameter4, KMockTypeParameter5, *>,
+    KMockTypeParameter0 : Any, KMockTypeParameter1, KMockTypeParameter2 : Any, KMockTypeParameter3,
+    KMockTypeParameter4, KMockTypeParameter5> kmock(
     collector: KMockContract.Collector = NoopCollector,
     relaxed: Boolean = false,
     relaxUnitFun: Boolean = false,
     freeze: Boolean = true,
-    templateType0: kotlin.reflect.KClass<multi.template.platformGeneric.Generic1<KMockTypeParameter0,
-        KMockTypeParameter1>>,
-    templateType1: kotlin.reflect.KClass<multi.template.platformGeneric.nested.Generic2<KMockTypeParameter2,
-        KMockTypeParameter3>>,
-    templateType2: kotlin.reflect.KClass<multi.template.platformGeneric.GenericPlatformContract.Generic3<KMockTypeParameter4,
-        KMockTypeParameter5>>,
-): Mock where Mock : Generic1<KMockTypeParameter0, KMockTypeParameter1>, Mock :
-Generic2<KMockTypeParameter2, KMockTypeParameter3>, Mock :
-              GenericPlatformContract.Generic3<KMockTypeParameter4, KMockTypeParameter5>, KMockTypeParameter1
-              : Any, KMockTypeParameter1 : Comparable<KMockTypeParameter1>, KMockTypeParameter3 : Any,
-              KMockTypeParameter3 : Comparable<KMockTypeParameter3> = getMockInstance(
-    spyOn = null,
-    collector = collector,
-    relaxed = relaxed,
-    relaxUnitFun = relaxUnitFun,
-    freeze = freeze,
-    templateType0 = templateType0,
-    templateType1 = templateType1,
-    templateType2 = templateType2,
-)
+    templateType0: KClass<Generic1<*, *>>,
+    templateType1: KClass<Generic2<*, *>>,
+    templateType2: KClass<GenericPlatformContract.Generic3<*, *>>,
+): Mock where KMockTypeParameter1 : Any, KMockTypeParameter1 : Comparable<KMockTypeParameter1>,
+              KMockTypeParameter3 : Any, KMockTypeParameter3 : Comparable<KMockTypeParameter3> =
+    getMockInstance(
+        spyOn = null,
+        collector = collector,
+        relaxed = relaxed,
+        relaxUnitFun = relaxUnitFun,
+        freeze = freeze,
+        templateType0 = templateType0,
+        templateType1 = templateType1,
+        templateType2 = templateType2,
+    )