+++ kmock-processor/src/test/resources/multi/expected/sharedGeneric/SpiedGenericActualFactory.kt
@@ -6,6 +6,7 @@ import kotlin.Any
 import kotlin.Boolean
 import kotlin.Comparable
 import kotlin.Suppress
+import kotlin.reflect.KClass
 import multi.template.sharedGeneric.Generic1
 import multi.template.sharedGeneric.GenericSharedContract
 import multi.template.sharedGeneric.nested.Generic2
@@ -42,22 +43,15 @@ private inline fun <reified Mock : SpyOn, reified SpyOn, KMockTypeParameter0 : A
     relaxed: Boolean,
     relaxUnitFun: Boolean,
     freeze: Boolean,
-    templateType0: kotlin.reflect.KClass<multi.template.sharedGeneric.Generic1<KMockTypeParameter0,
-        KMockTypeParameter1>>,
-    templateType1: kotlin.reflect.KClass<multi.template.sharedGeneric.nested.Generic2<KMockTypeParameter2,
-        KMockTypeParameter3>>,
-    templateType2: kotlin.reflect.KClass<multi.template.sharedGeneric.GenericSharedContract.Generic3<KMockTypeParameter4,
-        KMockTypeParameter5>>,
+    templateType0: KClass<Generic1<*, *>>,
+    templateType1: KClass<Generic2<*, *>>,
+    templateType2: KClass<GenericSharedContract.Generic3<*, *>>,
 ): Mock where SpyOn : Generic1<KMockTypeParameter0, KMockTypeParameter1>, SpyOn :
 Generic2<KMockTypeParameter2, KMockTypeParameter3>, SpyOn :
               GenericSharedContract.Generic3<KMockTypeParameter4, KMockTypeParameter5>, KMockTypeParameter1 :
               Any, KMockTypeParameter1 : Comparable<KMockTypeParameter1>, KMockTypeParameter3 : Any,
-              KMockTypeParameter3 : Comparable<KMockTypeParameter3> = if (Mock::class ==
-    multi.SharedGenericMultiMock::class) {
-    multi.SharedGenericMultiMock(collector = collector, freeze = freeze, spyOn = spyOn as SpyOn?) as
-        Mock
-} else {
-    throw RuntimeException("Unknown Interface ${Mock::class.simpleName}.")}
+              KMockTypeParameter3 : Comparable<KMockTypeParameter3> = multi.SharedGenericMultiMock(collector =
+collector, freeze = freeze, spyOn = spyOn as SpyOn?) as Mock
 
 internal actual inline fun <reified Mock : SpyOn, reified SpyOn, KMockTypeParameter0 : Any,
     KMockTypeParameter1, KMockTypeParameter2 : Any, KMockTypeParameter3, KMockTypeParameter4,
@@ -65,12 +59,9 @@ internal actual inline fun <reified Mock : SpyOn, reified SpyOn, KMockTypeParame
     spyOn: SpyOn,
     collector: KMockContract.Collector,
     freeze: Boolean,
-    templateType0: kotlin.reflect.KClass<multi.template.sharedGeneric.Generic1<KMockTypeParameter0,
-        KMockTypeParameter1>>,
-    templateType1: kotlin.reflect.KClass<multi.template.sharedGeneric.nested.Generic2<KMockTypeParameter2,
-        KMockTypeParameter3>>,
-    templateType2: kotlin.reflect.KClass<multi.template.sharedGeneric.GenericSharedContract.Generic3<KMockTypeParameter4,
-        KMockTypeParameter5>>,
+    templateType0: KClass<Generic1<*, *>>,
+    templateType1: KClass<Generic2<*, *>>,
+    templateType2: KClass<GenericSharedContract.Generic3<*, *>>,
 ): Mock where SpyOn : Generic1<KMockTypeParameter0, KMockTypeParameter1>, SpyOn :
 Generic2<KMockTypeParameter2, KMockTypeParameter3>, SpyOn :
               GenericSharedContract.Generic3<KMockTypeParameter4, KMockTypeParameter5>, KMockTypeParameter1 :