+++ kmock-processor/src/test/resources/multi/expected/sharedGeneric/SpiedGenericExpectFactory.kt
@@ -6,6 +6,7 @@ import kotlin.Any
 import kotlin.Boolean
 import kotlin.Comparable
 import kotlin.Suppress
+import kotlin.reflect.KClass
 import multi.template.sharedGeneric.Generic1
 import multi.template.sharedGeneric.GenericSharedContract
 import multi.template.sharedGeneric.nested.Generic2
@@ -19,12 +20,9 @@ internal expect inline fun <reified Mock : SpyOn, reified SpyOn, KMockTypeParame
     spyOn: SpyOn,
     collector: KMockContract.Collector = NoopCollector,
     freeze: Boolean = true,
-    templateType0: kotlin.reflect.KClass<multi.template.sharedGeneric.Generic1<KMockTypeParameter0,
-        KMockTypeParameter1>>,
-    templateType1: kotlin.reflect.KClass<multi.template.sharedGeneric.nested.Generic2<KMockTypeParameter2,
-        KMockTypeParameter3>>,
-    templateType2: kotlin.reflect.KClass<multi.template.sharedGeneric.GenericSharedContract.Generic3<KMockTypeParameter4,
-        KMockTypeParameter5>>,
+    templateType0: KClass<Generic1<*, *>>,
+    templateType1: KClass<Generic2<*, *>>,
+    templateType2: KClass<GenericSharedContract.Generic3<*, *>>,
 ): Mock where SpyOn : Generic1<KMockTypeParameter0, KMockTypeParameter1>, SpyOn :
 Generic2<KMockTypeParameter2, KMockTypeParameter3>, SpyOn :
               GenericSharedContract.Generic3<KMockTypeParameter4, KMockTypeParameter5>, KMockTypeParameter1 :